Uploaded all files from this project
This project is about selecting monsters from a list with weighted selection.
In this project, I add monsters to a list, and give each of them a weight.
When called, a monster's name is provided. The total number of times each monster appears when the program is run many times will result in the weightage allocated to each monster.